eventListener to fill in gridItem divs when click
Originally I was writing
    gridItem.addEventListener("mouseover", (e)=>
    gridItem.classList.add("someClass"));

However, it can't work because the gridItem variable doesn't have
global scope. Furthermore, I think it would need to somehow be applied
to each gridItem variable which could be difficult and ineffecient.

Instead I used .target to delegate the event to trigger when the parent
node (GRID_CONTAINER) was clicked/hovered over. Then retrieve the child
node that triggered the event within the parent node.

By adding the if(e.target.id) it would stop the parent div/node from
triggering the event.

I then added a classList to the node that triggered the event.

I read about the delegation here:
https://blog.bitsrc.io/cleaner-code-with-event-delegation-in-javascript-bb6dd27be64d

But I actually used a similar method when working on my
Rock Paper Scissors app.