Added button and eventListener and a function
Trying to add a feature to change the grid size.
User is supposed to click on a button and receive a prompt.
User enters a number and the grid size is redrawn.
Unfortunately, I'm not sure how to implement and will need to think
about how to do it.
Might need to refactor some earlier code.